Add user table

+++ db/migrate/01_create_stocks.rb
@@ -5,8 +5,8 @@ class CreateStocks < ActiveRecord::Migration
       t.integer :company_id
       t.datetime :open_date
       t.datetime :close_date
-      t.integer :open_price
-      t.integer :close_price
+      t.decimal :open_price
+      t.decimal :close_price
       t.integer :percentage_change
     end
   end
